requirements.txt
annotated-types==0.5.0
anyio==3.7.1
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.6
colorama==0.4.6
decorator==5.1.1
einops==0.6.1
encodec==0.1.1
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.9.0
freetype-py==2.4.0
fsspec==2023.6.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.1
librosa==0.8.1
llvmlite==0.40.1
MarkupSafe==2.1.2
miniaudio==1.59
mpmath==1.2.1
networkx==3.0
numba==0.57.1
numpy==1.22.4
open3d==0.17.0
opencv-python==4.8.0.74
packaging==23.1
Pillow==9.3.0
platformdirs==3.9.1
pooch==1.7.0
protobuf==4.23.4
psbody-mesh @ git+https://github.com/a-elrawy/mesh.git@a3f52f57b6c7f97b84b6505a0f10106f822a6baf
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pyglet==1.5.27
Pygments==2.16.1
PyOpenGL==3.1.0
pyrender==0.1.45
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.1
resampy==0.4.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.7.2
six==1.16.0
smplx==0.1.28
sniffio==1.3.0
soundfile==0.12.1
starlette==0.27.0
sympy==1.11.1
tensorboardX==2.6.1
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
trimesh==3.9.27
typing_extensions==4.7.1
urllib3==1.26.13
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
wandb
google-cloud-storage